Create default config files manually
Sometimes `rules_clojure` fails with an error in
`tools_deps.bzl/_run_gen_build`, which claims that it couldn't copy to
`~/.clojure/deps.edn`, because the file already exists.

This appears to be a step in the `clojure` script. When run, it checks
for the existence of `~/.clojure/deps.edn` (by default), and copies a
file into there from the install dir if not.

My theory is that multiple `clojure` runs are occurring at the same time
in the same container. When that happens, there can be a race condition
where they both detect that the file is not present, and then both try
to copy the example copy in.

To prevent this race condition due to the behaviour of `cp`, we now set
up these files within the `rules_clojure` Bazel functions, before first
calling `clojure`.