Use dependency pinning
Some minor refactors, for consistency:
- Add `.cpcache` to `.gitignore`, so you can run `bazel` commands
  in-repo and not get spurious files.
- Rename `BUILD.bazel` files to `BUILD`, as that is the newer pattern,
  and we had a mixture of both.

And now, the main event:

We're now following the pattern that the Bazel team themselves use for
dependencies of rules, particularly for JVM-based languages making use
of `rules_jvm_external`. E.g. this is used in `contrib/rules_jvm`, among
others.

- We've pinned the dependencies for `rules_clojure` in
  `frozen_deps_install.json`, using standard options built into
`maven_install`.
- There's a new `zip_repository` definition, that extracts the files
  necessary for Bazel to fetch dependencies from a zipfile.
- This `zip_repository` is what we reference everywhere when referring
  to `rules_clojure`'s internal dependencies, so is what is run by users
of `rules_clojure` to set it up.
- To construct the zipfile, there is a (for now, Python) script under
  `tools/freeze-deps.py`. Instructions in `README.md` explain how to
update deps when necessary.

The Python script is pretty simple, and could be changed to a Clojure
one in future. For now, while proving the idea, I've kept it as is. Both
it and `zip_repository.bzl` were taken from `contrib/rules_jvm`, where
they're licensed under Apache 2. One of the devs of `rules_jvm_external`
and that (seems to do most of the releases for both) helped me out with
all this and pointed me at them.