Fix `persistent-worker` exit code
According to the docs at
https://bazel.build/remote/persistent#implementation, the protobuf
definition for persistent workers uses snake case, but the JSON
implementation uses camel case. It also specifies that the JSON version
should ignore any fields it gets that it doesn't recognise.

Given `:exit_code` was specified in our response map, instead of
`:exitCode`, I believe Bazel was never detecting the exit code from a
persistent worker job, and was just defaulting to the assumption that
the build succeeded. This is why the compilation failure in the code the
data team is working on for the chart of accounts migration etc managed
to get into the primary branch.